Revert "fixed compiler errors"
This reverts commit 1fa695d91551b71e41ff7914d31f37859865c8a7.

+++ FtcRobotController/src/main/java/com/qualcomm/hardware/MatrixDcMotorController.java
@@ -72,7 +72,7 @@ public class MatrixDcMotorController implements DcMotorController {
         this.f91b = 0;
         master.registerMotorController(this);
         for (int i = 0; i < 4; i = (byte) (i + 1)) {
-            master.queueTransaction(new MatrixI2cTransaction((byte) i, (byte) 0, 0, (byte) 0));
+            master.queueTransaction(new MatrixI2cTransaction(i, (byte) 0, 0, (byte) 0));
             this.f90a[i].f88f = RunMode.RUN_WITHOUT_ENCODERS;
             this.f90a[i].f86d = true;
         }
